services/ComputeServer/requirements.txt
celery
python-dotenv
transformers
numpy
boto3
minio
sqlalchemy
psycopg2-binary
# opencv-python
opencv-python-headless
pillow
pathlib
scikit-image
requests
redis
pydantic
openai
google-generativeai
torch
torchvision
--extra-index-url https://download.pytorch.org/whl/cu126
sentence-transformers
jieba
rank_bm25
pgvector